In the middle of cart_items
Getting this to work required some migrations, and apparently, there
was an issue with some duplicated migrations. I added an 'open_order'
attribute to the users model, which should resolve the problem of
creating orders. I still have to make this work for automatically
created orders via selecting an item (which is converted to a
cart_item, but that first needs an order to be created).