Cart_items
They are getting the current order now. Planning out how I want to
display the prices on the cart_item form. I don't think I want to.
Psychological reasons are behind this. The price will only be displayed
at the end. I am most likely not going to be using the price attribute
of cart items, as that is potentially insecure.