Basic data structure modification
It works, and it will potentially support modifiers moving between
modifier groups.
<br>
Whoa. So, I have meats and veggies as the LIST modifier groups...
But the modifier groups that the kitchen gets are WHOLE LEFT RIGHT.
The same modifier will not ever be in more than one group at a time.
<br>
Still need to put modifiers on modifiers to set amounts (light,
regular, double). But this should be abstracted in much the same way as
the sizes are done... That way I can support sauce settings and cheese
settings. Each modifier_modifier needs a name and a price.