Processing Port p1
porting my processing pizza topping algorithm. This has so far been a
pretty tedious process, with a lots of trial and error. What I've found
is that it's actually a bit more straight-forward than I thought for
handling these non-draggable elements, because of where I've put them
in the code (before all the columns and stuff that made the draggables
so darn complicated).